refactor(frontend): reduce over-decomposition in event details
