+++ frontend/src/features/events/ui/event-details/EventCard.stories.tsx
@@ -48,18 +48,14 @@ const meta = {
   },
   args: {
     event: baseEvent,
-    state: {
-      token: "token",
-      isOrganizer: false,
-      isJoined: false,
-      isBusy: false,
-    },
-    handlers: {
-      onOpen: () => undefined,
-      onJoin: () => undefined,
-      onLeave: () => undefined,
-      onRequireLogin: () => undefined,
-    },
+    token: "token",
+    isOrganizer: false,
+    isJoined: false,
+    isBusy: false,
+    onOpen: () => undefined,
+    onJoin: () => undefined,
+    onLeave: () => undefined,
+    onRequireLogin: () => undefined,
   },
 } satisfies Meta<typeof EventCard>;
 
@@ -71,23 +67,19 @@ export const Default: Story = {};
 
 export const Joined: Story = {
   args: {
-    state: {
-      token: "token",
-      isOrganizer: false,
-      isJoined: true,
-      isBusy: false,
-    },
+    token: "token",
+    isOrganizer: false,
+    isJoined: true,
+    isBusy: false,
   },
 };
 
 export const OrganizerView: Story = {
   args: {
-    state: {
-      token: "token",
-      isOrganizer: true,
-      isJoined: false,
-      isBusy: false,
-    },
+    token: "token",
+    isOrganizer: true,
+    isJoined: false,
+    isBusy: false,
   },
 };
 
@@ -111,11 +103,9 @@ export const FullEvent: Story = {
         },
       ],
     },
-    state: {
-      token: "token",
-      isOrganizer: false,
-      isJoined: false,
-      isBusy: false,
-    },
+    token: "token",
+    isOrganizer: false,
+    isJoined: false,
+    isBusy: false,
   },
 };
