+++ frontend/src/features/events/ui/event-details/EventCardGrid.tsx
@@ -40,20 +40,17 @@ export function EventCardGrid({ events }: EventCardGridProps) {
           <EventCard
             key={event.id}
             event={event}
-            state={{
-              token,
-              isOrganizer:
-                Boolean(currentUserEmail) &&
-                currentUserEmail === event.organizer?.email,
-              isJoined: joinedEventIds.has(event.id),
-              isBusy: busyEventId === event.id,
-            }}
-            handlers={{
-              onOpen: () => navigate(`/events/${event.id}`),
-              onJoin: () => void handleJoin(event.id),
-              onLeave: () => void handleLeave(event.id),
-              onRequireLogin: () => navigate("/login"),
-            }}
+            token={token}
+            isOrganizer={
+              Boolean(currentUserEmail) &&
+              currentUserEmail === event.organizer?.email
+            }
+            isJoined={joinedEventIds.has(event.id)}
+            isBusy={busyEventId === event.id}
+            onOpen={() => navigate(`/events/${event.id}`)}
+            onJoin={() => void handleJoin(event.id)}
+            onLeave={() => void handleLeave(event.id)}
+            onRequireLogin={() => navigate("/login")}
           />
         ))}
       </div>
